pyproject.toml
[project]
name = "comfyui-replenishnodes"
description = "The nodes in this repository are only used as secondary nodes."
version = "1.1.0"
license = {file = "LICENSE"}
dependencies = ["numpy", "opencv-python", "Pillow"]

[project.urls]
Repository = "https://github.com/hben35096/ComfyUI-ReplenishNodes"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = ""
DisplayName = "ComfyUI-ReplenishNodes"
Icon = ""
